🚚 [Move] timer 위젯별 파일 분리
