+++ lib/ui/todo/widget/timer_log/linear_timer_log.dart
@@ -1,4 +1,4 @@
-import 'package:time_todo/ui/utils/timer_log_entry.dart';
+import 'package:time_todo/ui/todo/widget/timer_log/timer_log_entry.dart';
 
 class LinearTimerLog {
   // 지난 로그들을 그리기 위해 저장할 변수
