+++ lib/ui/todo/widget/timer_log/timer_log_list_header.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:time_todo/assets/colors/color.dart';
 import 'package:time_todo/ui/components/widget/app_components.dart';
-import 'package:time_todo/ui/todo/widget/timer_log_list.dart';
-import '../../utils/timer_log_entry.dart';
+import 'package:time_todo/ui/todo/widget/timer_log/timer_log_entry.dart';
+import 'package:time_todo/ui/todo/widget/timer_log/timer_log_list.dart';
 
 class TimerLogListHeader extends StatefulWidget {
   final List<TimerLogEntry> timerLog;
